renames navigation test to navigation router
