+++ App.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { Provider } from 'react-redux';
 import Dashboard from './src/screens/dashboard';
 import store from './src/store';
-import TabNavigator from './src/navigation/navigationtest'
+import TabNavigator from './src/navigation/navigationRouter'
